requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.6
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
flower==2.0.1
humanize==4.14.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
paystackapi==2.1.3
phonenumbers==9.0.15
pillow==12.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
